ActivityContainer gefixt
Long en Integer meuk

+++ Wielwijk/src/wielwijk/ActivityContainer.java
@@ -51,7 +51,7 @@ public class ActivityContainer {
         
         return new Activity((String) map.get("name"), (String) map.get("location"), (String) map.get("destination"), datetimebegin, datetimeend,
                 (Integer) map.get("fee"), (Integer) map.get("lower_user_limit"), (Integer) map.get("upper_user_limit"),
-                map.get("lower_limit_date").toString(), (Boolean) map.get("cancelled"), (Integer) map.get("id"));
+                map.get("lower_limit_date").toString(), (Boolean) map.get("cancelled"), ((Long) map.get("id")).intValue());
     }
 
     public static Activity getActivityByDay(int year, int month, int day) {
@@ -72,6 +72,6 @@ public class ActivityContainer {
 
         return new Activity((String) map.get("name"), (String) map.get("location"), (String) map.get("destination"), datetimebegin, datetimeend,
                 (Integer) map.get("fee"), (Integer) map.get("lower_user_limit"), (Integer) map.get("upper_user_limit"),
-                map.get("lower_limit_date").toString(), (Boolean) map.get("cancelled"), (Integer) map.get("id"));
+                map.get("lower_limit_date").toString(), (Boolean) map.get("cancelled"), ((Long) map.get("id")).intValue());
     }
 }
